refactor: create modules folder, move files into it
