+++ src/modal.ts
@@ -1,5 +1,5 @@
 /* ===============================
-  MODAL & CLEAR functionality for textarea
+  MODAL functionality for textarea
 ================================== */
 const textBox = document.getElementById('text-box') as HTMLTextAreaElement;
 const modal = document.getElementById("modal")! as HTMLDivElement;
@@ -15,10 +15,3 @@ export function closeModalOnClick(): void {
 }
 
 window.addEventListener("click", e => (e.target === modal ? modal.classList.remove("show-modal") : false));
-
-// clear textarea
-export function clearTextarea(): void {
-  textBox.value = "";
-  textBox.focus();
-  modal.classList.remove("show-modal");
-}
